requirements.txt
absl-py==2.1.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==16.1.0
awscli==1.40.3
Babel==2.14.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.41.0
bleach==6.2.0
blinker==1.4
blis==0.7.11
boto3==1.38.4
botocore==1.38.4
bottle==0.13.3
brainpp==2.7.12.12
cachetools==5.5.2
catalogue==2.0.10
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
cloud-init==24.4.1
cloudpathlib==0.23.0
colorama==0.4.4
comm==0.2.2
confection==0.1.5
configobj==5.0.6
confluent-kafka==2.10.0
contourpy==1.3.2
cryptography==3.4.8
cycler==0.12.1
cymem==2.0.13
Cython==0.29.28
datasets==2.16.1
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.14.4
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
docstring_parser==0.17.0
docutils==0.19
dotmap==1.3.30
durationpy==0.9
einops==0.6.1
einops-exts==0.0.4
elasticsearch==7.17.12
en-core-web-sm @ file:///mnt/shared-storage-user/steai-share/gaozhenkun/syx_code/ZoomEye-main/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.125.0
fastjsonschema==2.21.1
filelock==3.20.2
Flask==2.2.5
fonttools==4.57.0
frozendict==1.2
frozenlist==1.8.0
fsspec==2023.10.0
ftfy==6.3.1
future==0.18.2
gitdb==4.0.12
gitignore-parser==0.1.12
GitPython==3.1.46
google-auth==2.39.0
gradio_client==0.2.9
graphviz==0.20.3
h11==0.14.0
h5py==3.13.0
hf-xet==1.2.1
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.0
huggingface-hub==0.36.0
idna==3.3
ImageIO==2.37.2
importlib-metadata==4.6.4
influxdb==5.3.2
ipykernel==6.29.5
ipython==8.36.0
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
json5==0.12.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-server==1.24.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab==3.3.2
jupyterlab-drawio==0.9.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
kiwisolver==1.4.8
kubernetes==32.0.1
langcodes==3.5.1
latex2mathml==3.78.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
-e git+https://github.com/xiaoxiang100/training-free-work.git@594f5ad847c260dfb5b0bddbe60f798a8197105e#egg=llava&subdirectory=ZoomEye-main-12.18/ZoomEye-main
lmdb==1.6.2
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
megfile==2.2.10.post1
mistune==3.1.3
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.7.0
multiprocess==0.70.15
murmurhash==1.0.15
nbclassic==0.5.6
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
ninja==1.13.0
notebook_shim==0.2.4
numpy==1.26.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.590.44
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.46
open_clip_torch==3.2.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
peft==0.4.0
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.3.7
preshed==3.0.12
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==6.33.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.8
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
PyOpenGL==3.1.9
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-distutils-extra==2.39
pytz==2022.1
PyYAML==5.4.1
pyzmq==26.4.0
qwen-vl-utils==0.0.14
redis==5.2.1
referencing==0.36.2
refile==7.2.7.post3
regex==2025.11.3
requests==2.32.3
requests-futures==1.0.0
requests-oauthlib==2.0.0
retry==0.9.2
rich==14.2.0
roman==3.3
rpds-py==0.24.0
rsa==4.7.2
s3transfer==0.12.0
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
sentencepiece==0.1.99
sentry-sdk==2.49.0
setproctitle==1.3.5
shapely==2.1.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.50.0
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.24
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
triton==3.2.0
typeguard==4.4.2
typer==0.21.1
typer-slim==0.21.1
typing_extensions==4.13.2
tyro==1.0.5
tzdata==2025.2
urllib3==2.0.0
uvicorn==0.40.0
wadllib==1.3.6
waitress==3.0.2
wandb==0.23.1
wasabi==1.1.3
watchdog==2.1.6
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.3
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.6.0
yarl==1.22.0
zipp==1.0.0
